docs: QwoaN improvements - navigation and error handling
Holistic improvements based on Christopher Alexander's "Quality Without a Name":

## Navigation (Gradient - slope not cliff)
- Add "Start Here" section to root CLAUDE.md for new inhabitants
- Include "Where Things Go" table for common patterns
- Add "Key Naming Conventions" (Effect vs Effects, Graph.Tool)
- Add "See Also" sections to Tool modules for cross-navigation:
  - Tidepool.Graph.Tool → Wire, Convert, Schema
  - Tidepool.Tool.Wire → Graph.Tool, Convert, Schema
  - Tidepool.Tool.Convert → Graph.Tool, Wire, Schema
  - Tidepool.LLM.Types → LLMProvider, Wire

## Error Handling (Principled Library Pattern)
- Move LLMError type to Tidepool.Effects.LLMProvider (single source)
- Add completeTry variant returning Either LLMError (Try pattern)
- Add CompleteTry constructor to LLMComplete effect
- Update executor to handle both Complete and CompleteTry
- Re-export LLMError from LLM.Types for backwards compatibility
- Document throwing vs try variants in Haddock

This follows the "Effect + Try variants" pattern already established
in Tidepool.Effects.Telegram and other effect modules.
