+++ tidepool-core/src/Tidepool/Graph/Tool.hs
@@ -9,6 +9,12 @@
 -- * The tool marker type @t@ is passed to methods (can carry config)
 -- * Effect stacks are declared as associated types
 --
+-- = See Also
+--
+-- * "Tidepool.Tool.Wire" - Wire format types for Anthropic and OpenAI APIs
+-- * "Tidepool.Tool.Convert" - Conversion typeclasses ('ToAnthropicTool', 'ToCfTool')
+-- * "Tidepool.Schema" - JSON Schema derivation for tool input types
+--
 -- = Design Principle
 --
 -- Everything hangs off the 'ToolDef' typeclass:
