+++ tidepool-native-gui/llm-executor/src/Tidepool/LLM/Types.hs
@@ -2,6 +2,11 @@
 --
 -- Provides configuration for Anthropic and OpenAI API clients.
 --
+-- = See Also
+--
+-- * "Tidepool.Effects.LLMProvider" - Effect definition and 'LLMError' type
+-- * "Tidepool.Tool.Wire" - Wire format types for tools
+--
 -- = Tool Schema Formats
 --
 -- Anthropic and OpenAI use different tool definition formats:
@@ -35,7 +40,7 @@ module Tidepool.LLM.Types
   , LLMEnv(..)
   , mkLLMEnv
 
-    -- * Error Types
+    -- * Error Types (re-exported from "Tidepool.Effects.LLMProvider")
   , LLMError(..)
 
     -- * Tool Definitions (re-exported from "Tidepool.Tool.Wire")
@@ -47,6 +52,9 @@ import Data.Text (Text)
 import GHC.Generics (Generic)
 import Network.HTTP.Client (Manager, newManager)
 import Network.HTTP.Client.TLS (tlsManagerSettings)
+
+-- Re-export error types from effect module
+import Tidepool.Effects.LLMProvider (LLMError(..))
 
 -- Re-export wire types
 import Tidepool.Tool.Wire (AnthropicTool(..), anthropicToolToJSON)
@@ -102,22 +110,5 @@ mkLLMEnv :: LLMConfig -> IO LLMEnv
 mkLLMEnv config = do
   manager <- newManager tlsManagerSettings
   pure LLMEnv { leConfig = config, leManager = manager }
-
-
--- ════════════════════════════════════════════════════════════════════════════
--- ERROR TYPES
--- ════════════════════════════════════════════════════════════════════════════
-
--- | LLM API errors.
-data LLMError
-  = LLMHttpError Text           -- ^ Network/HTTP error
-  | LLMParseError Text          -- ^ JSON parsing failed
-  | LLMRateLimited              -- ^ Rate limit hit (429)
-  | LLMUnauthorized             -- ^ Invalid API key (401)
-  | LLMContextTooLong           -- ^ Input too long (400 with specific error)
-  | LLMOverloaded               -- ^ Service overloaded (529)
-  | LLMApiError Text Text       -- ^ API error (type, message)
-  | LLMNoProviderConfigured     -- ^ No provider secrets configured
-  deriving stock (Eq, Show, Generic)
